Renamed MacroModel.jl to Macro (and module)
This is more similar to other naming conventions in LEAP-Macro.
